feat: refactor the package of credentials model classes
