+++ .gitignore
@@ -6,3 +6,5 @@ build
 /foobar
 /gradle.properties
 target
+.DS_Store
+apikit-flow-generator-maven-plugin.iml
