+++ bin/src/main/groovy/com/avioconsulting/mule/designcenter/HttpClientWrapper.groovy
