+++ bin/src/main/groovy/com/avioconsulting/mule/designcenter/api/model/credentials/ConnectedAppCredential.groovy
