+++ src/main/groovy/com/avioconsulting/mule/anypoint/api/credentials/model/Credential.groovy
@@ -1,0 +1,9 @@
+package com.avioconsulting.mule.anypoint.api.credentials.model;
+
+/**
+ * Abstract Base class for defining anypoint platform access credentials
+ */
+abstract class Credential {
+
+    abstract String getPrincipal()
+}
